bot/requirements.txt
discord.py~=2.6.4
lavalink~=5.9.0
pydantic~=2.12.5
pydantic-settings~=2.12.0
pymongo~=4.16.0
python-dotenv~=1.2.1
